moving types to types.ts
