+++ nexus-sdk-js/src/Organization/index.ts
@@ -6,37 +6,8 @@ import {
   deprecateOrganization,
   updateOrganization,
 } from './utils';
-import { Context, CreateOrgPayload } from './types';
+import { Context, CreateOrgPayload, OrgResponse } from './types';
 import { ListProjectOptions, CreateProjectPayload } from '../Project/types';
-
-export interface OrgResponseCommon {
-  '@id': string;
-  '@type': string;
-  _uuid: string;
-  _label: string;
-  _rev: number;
-  _deprecated: boolean;
-  _createdAt: string;
-  _createdBy: string;
-  _updatedAt: string;
-  _updatedBy: string;
-  _self?: string;
-  _constrainedBy?: string;
-  description?: string;
-}
-
-export interface ListOrgResponse {
-  '@context': Context;
-  _total: number;
-  _results?: OrgResponseCommon[];
-  _links?: any;
-  code?: string;
-  message?: string;
-}
-
-export interface OrgResponse extends OrgResponseCommon {
-  '@context': Context;
-}
 
 export default class Organization {
   context?: Context;
