+++ nexus-sdk-js/src/Organization/types.ts
@@ -1,3 +1,32 @@
+export interface OrgResponseCommon {
+  '@id': string;
+  '@type': string;
+  _uuid: string;
+  _label: string;
+  _rev: number;
+  _deprecated: boolean;
+  _createdAt: string;
+  _createdBy: string;
+  _updatedAt: string;
+  _updatedBy: string;
+  _self?: string;
+  _constrainedBy?: string;
+  description?: string;
+}
+
+export interface ListOrgResponse {
+  '@context': Context;
+  _total: number;
+  _results?: OrgResponseCommon[];
+  _links?: any;
+  code?: string;
+  message?: string;
+}
+
+export interface OrgResponse extends OrgResponseCommon {
+  '@context': Context;
+}
+
 export interface ListOrgOptions {
   full_text_search?: string;
   from?: number;
